Bioshake

Description

Bioshake is a small framework for specifying bioinformatics pipelines. The goal is to specify phases in a forward chaining manner (as is natural for the domain) while guaranteeing as much robustness as possible to errors such as mismatched file types or other attributes. Almost everything is handled in the type system, and pipelines are compiled down to Development.Shake Rules for actual execution.

Documentation

data All a where

Datatype to represent fan-in combinations.

Constructors

All :: (Functor f, Foldable f) => f a -> All a

class Referenced a where

Attaches a reference genome.

Minimal complete definition

getRef, name

Methods

getRef :: a -> FilePath

The path to the reference fasta file.

name :: a -> String

The short name, e.g., hg19.

dbnsfp :: a -> FilePath

Path to dbNSFP for the genome

class Capture a where

Asserts a capture region.

Minimal complete definition

getBED

Methods

getBED :: a -> FilePath

ignoringIOErrors :: IO () -> IO ()

withTempDirectory

Arguments

:: FilePath
Target directory under which the temporary directory is created

-> String
Template for the temporary directory name

-> (FilePath -> Action b)
Action to carry out

-> Action b

Creates a temporary directory under a target directory according to a naming template. The directory is cleaned up after executing the action. This differs from Development.Shake's withTempDir in that it takes a target directory and template whereas Development.Shake uses /tmp. This is generally more useful, as ./tmp is used as the target directory by convention in BioShake.

bioshake

Arguments

:: Int
Number of threads

-> ShakeOptions
Options to pass to shakeArgs.

-> (Implicit Resource => Rules ())

-> IO ()

Entry point to bioshake. Like shakeArgs but also takes a number of threads to use.

out :: [FilePath] -> Out

Explicitly names an output product. Outputs are automatically named in the temporary directory except for this special case: this is how you obtain the artifacts you are specifically interested in.

withAll :: (Functor f, Foldable f) => f a -> All a

Fan-in style combinator. Takes a collection of combines their output paths as input paths for the subsequent phase.

withPair :: a -> a -> All a

Explicitly construct a fan-in of exactly two items