Bioshake.Tags

Description

This module introduces a set of common tags that are frequently used when defining phases.

Documentation

class DeDuped a

Duplicated reads have been removed

class HasRG a

Contains an RG line

class PairedEnd a

Results from a paired end sequencing processed

class Sorted a

Sorted (e.g., sorted sam/bam, or bed)

class IsBam a

class IsBcf a

class IsBed a

class IsCSV a

class IsFastQ a

class IsGff a

class IsMPileup a

class IsSam a

class IsSeqzGZ a

class IsTSV a

class IsVCF a

class IsTGZ a

class IsCov a

allTags :: [Name]

allTransTagsPipe :: Name -> Q [Dec]

Template haskell to declare transitivity for all tags through a pipe ending in a type. Examples include Out.

allTransTags :: Name -> Q [Dec]

Same as allTransTagsPipe but for transitivity through a constructor. Examples include All.