Bioshake.Types

Description

Core data types the pipeline abstraction.

Documentation

data a :-> b where infixl 1

Pipes output of phase a into phase b, i.e., forms a pipeline.

Constructors

(:->) :: a -> b -> a :-> b infixl 1

class Buildable a where

Buildable abstracts things that can be turned into shake Actions.

Minimal complete definition

build

Methods

build :: Implicit Resource => a -> Action ()

type Compiler = StateT (Set [FilePath]) Rules

The compiler tracks the set of output files to ensure duplicate Rules are not generated. This allows multiple potentially overlapping pipelines to be compiled down to a set of unique Rules.

compileRules :: Compiler () -> Rules ()

Compile pipelines to Rules.

class Compilable a where

Pipelines are Compilable when they can be compiled down to a set of Rules that build a list of output paths.

Methods

compile :: Implicit Resource => a -> Compiler ()

class Pathable a where

Things are pathable if they can be mapped to a list of file paths. This is used to make the files for a phase concrete in the build system.

Minimal complete definition

paths

Methods

paths :: a -> [FilePath]

data Nuc

Constructors

A

C

G

T

newtype Seq

Constructors

Seq [Nuc]

data Threads

Constructors

Threads Int